improvement(requirements.txt): bumped pillow to 6.2.2 to avoid DoS vulnerability
There is a DoS vulnerability in Pillow before 6.2.2 caused by FpxImagePlugin.py calling the range
function on an unvalidated 32-bit integer if the number of bands is large. On Windows running 32-bit
Python, this results in an OverflowError or MemoryError due to the 2 GB limit. However, on Linux
running 64-bit Python this results in the process being terminated by the OOM killer.